weights cannot be larger than uint

+++ fulqrum/core/qubit_operator.pyx
@@ -385,7 +385,7 @@ cdef class QubitOperator():
         Returns:
             ndarray: Array of operator weights
         """
-        cdef size_t[::1] out = np.zeros(self.oper.terms.size(), dtype=np.uint)
+        cdef unsigned int[::1] out = np.zeros(self.oper.terms.size(), dtype=np.uint32)
         cdef size_t kk
         for kk in range(self.oper.terms.size()):
             out[kk] = self.oper.terms[kk].values.size()
@@ -398,7 +398,7 @@ cdef class QubitOperator():
         Returns:
             ndarray: Array of operator off-diagonal weights
         """
-        cdef size_t[::1] out = np.zeros(self.oper.terms.size(), dtype=np.uint)
+        cdef unsigned int[::1] out = np.zeros(self.oper.terms.size(), dtype=np.uint32)
         cdef size_t kk
         for kk in range(self.oper.terms.size()):
             out[kk] = self.oper.terms[kk].offdiag_weight
